[REF] web_tour: move doAction/error to TourStep, factor macro steps on TourStepAutomatic
TourStep now owns doAction() and the error diagnostic (renamed from
describeWhyIFailed). TourAutomatic's per-step macro descriptors (log/debug
pair + trigger/action pair) move to a new TourStepAutomatic.actions getter.
